fix(desktop,trpc): address review comments on v2 workspaces filters PR
- Fix @superset/trpc test: add `verifyOrgMembershipWithSubscription` to
  the `../integration/utils` module mock so bun can resolve the named
  export pulled in transitively via `requireActiveOrgMembership`. The
  paywall PR added that export but the test mock wasn't updated, which
  was failing CI on main.
- Drop the auto-stamp narrative comment on github_pull_requests.updatedAt.
- Don't blank the list while machineId is unresolved: treat
  "this-device" as no-op until machineId is known.
- Auto-expand a project section when one of its workspaces is the
  active route, even if the user previously persisted it as collapsed.
- Split the inline helper components out of V2WorkspacePrHoverCardContent
  and V2WorkspacesHeader into one-component-per-folder modules per
  AGENTS.md.
