+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspaces/components/V2WorkspacePrHoverCardContent/components/PrStateBadge/index.ts
@@ -1,0 +1,1 @@
+export { PrStateBadge } from "./PrStateBadge";
