+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspaces/components/V2WorkspacesHeader/components/DeviceOptionLabel/index.ts
@@ -1,0 +1,1 @@
+export { DeviceOptionLabel } from "./DeviceOptionLabel";
