+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspaces/components/V2WorkspacesHeader/components/ProjectFilterTriggerLabel/index.ts
@@ -1,0 +1,1 @@
+export { ProjectFilterTriggerLabel } from "./ProjectFilterTriggerLabel";
